[examples] Fix rod2d inline style violations
This is exclusively code moving; no functional changes.
